Rename prelude to DSL.
